Move Checked<Tx> type to the fuel-vm crate (#310)
