+++ fuel-tx/src/transaction/types.rs
@@ -7,14 +7,10 @@ mod storage;
 mod utxo_id;
 mod witness;
 
-#[cfg(feature = "std")]
-pub use create::checked::CheckedMetadata as CreateCheckedMetadata;
 pub use create::Create;
 pub use input::{Input, InputRepr};
 pub use mint::Mint;
 pub use output::{Output, OutputRepr};
-#[cfg(feature = "std")]
-pub use script::checked::CheckedMetadata as ScriptCheckedMetadata;
 pub use script::Script;
 pub use storage::StorageSlot;
 pub use utxo_id::UtxoId;
