+++ fuel-vm/Cargo.toml
@@ -14,6 +14,7 @@ description = "FuelVM interpreter."
 [dependencies]
 anyhow = { version = "1.0", optional = true }
 backtrace = { version = "0.3", optional = true } # requires debug symbols to work
+bitflags = "1"
 dyn-clone = { version = "1.0", optional = true }
 fuel-asm = { workspace = true, path = "../fuel-asm" }
 fuel-crypto = { workspace = true, path = "../fuel-crypto" }
