+++ fuel-vm/tests/predicate.rs
@@ -182,8 +182,6 @@ fn predicate_gas_metering() {
 
 #[test]
 fn gas_used_by_predicates_is_deducted_from_script_gas() {
-    use fuel_vm::predicate::CheckedTxExt;
-
     let rng = &mut StdRng::seed_from_u64(2322u64);
 
     let gas_price = 1_000;
